Fix A*, ObjectParticle
A* 알고리즘 청크 이동 수정
오브젝트 잔해물 스크립트 수정

+++ Minecraft_Imitation/Assets/02.Scripts/NewBehaviourScript1.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class NewBehaviourScript1 : MonoBehaviour
 {
     public int g;
-    public int h;
+    public float h;
 }
